bugfix: Use index to get place name from all rows
The new Antelope row grabber doesn't use the record spec, it grabs
every row in the pointer, so we can either subset the table, or just
take all the rows and use the index to get the right one. An alternate
solution would be to subset the db, but we get all the rows, in case a
future implementation might want to do the distance/azimuth calc outside
of Antelope, in which case all the rows would be needed anyway.